requirements.txt
streamlit
pandas
requests
beautifulsoup4
python-dotenv
openai
crewai
langchain
langchain-openai
langchain-community
